requirements.txt
django==5.0
filetype==1.2.0
Pillow==10.1.0
